Move the normalizer static extensions to their own file
